fix(frontend): 修正 eslint 10.9.1 新提為 error 的兩條規則
eslint 10.9.1 把 react-hooks/set-state-in-effect 和
react-refresh/only-export-components 從 warning 提為 error，lint 從
baseline 的 0 errors / 42 warnings 變成 1 error 擋住 CI。

- ColorModeContext: 原本用 effect 在「使用者沒存過偏好」時把 mode
  同步成 OS 主題，這是典型的 setState-in-effect 級聯渲染。改成只把
  使用者明確選過的偏好放進 state，沒選過就直接由 prefersDark 推導，
  行為不變但少一次 render pass
- OverviewCard: RANK_COLORS 在該檔內部根本沒用到，只有 Rankings/index
  在用。搬到唯一使用者那邊，元件檔回復成純元件匯出，不需要多開一個
  常數檔
