+++ frontend/src/pages/Rankings/OverviewCard.jsx
@@ -1,12 +1,5 @@
 import { Card, CardActionArea, CardContent, Typography, Box } from "@mui/material";
 
-const RANK_COLORS = {
-  level: "#7c4dff",
-  gacha: "#ff6d00",
-  godStone: "#00bfa5",
-  achievement: "#d81b60",
-};
-
 export default function OverviewCard({ icon, title, topName, topValue, count, color, onClick }) {
   return (
     <Card
@@ -50,5 +43,3 @@ export default function OverviewCard({ icon, title, topName, topValue, count, co
     </Card>
   );
 }
-
-export { RANK_COLORS };
